Rearrange to better support collections of items
The App class will handle constructing and managing the model viewer
page. Collections of items go in src/collections, with a folder for each
one. So the Neyland stadium connection now lives in
src/collections/double-angle. The calculations, 2d drawings, etc. will
go in this folder as well when they're done.

New collections will need to be added as entry points in the webpack
config, and as output HTML files.