Add a working version of the sidebar
Some big changes here.

1. Website directory structure is changed
2. Header nav CSS is completely overhauled
3. Sizing of the viewer container is changed
4. The sidebar itself is added

1. Changes to directory structure
---------------------------------

To support having multiple files per model, the viewer app is now at
index.html inside a directory. This directory must have the same name as
the source directory in src/collections.

2. Header nav CSS
-----------------

To support the sidebar toggle, the header CSS has been completely
overhauled to use flexbox. An <h1> element for the page title has been
added; this is filled in by app.js. This is not final; there remain
some issues:

- The home icon appears to not be recognized as actually taking up
  space
- The title is intended to be flush left instead of centered
- Some issues can occur with titles being too long. It may or may not be
  worth the effort to determine how to automatically resize the text
  based on the available horizontal space (vertical resizing is already
  done).

3. Viewer container sizing
--------------------------

Since there's now the possibility of another element in the same row as
the viewer, the viewer will now calculate its width depending on the
state of the sidebar.

4. Sidebar: intent and next steps
---------------------------------

Ok. The sidebar itself.

This required a lot of development of the App class itself, and
rethinking the interface. The intent is for it to be *very* easy for
someone who has minimal knowledge of JavaScript/TypeScript/etc. to
contribute a model, with support pages and so forth.

Based on the mockup, the sidebar is broken up into groups. These groups
have a title, a list of items, and a <hr> between groups. The items may
be (at the moment) one of two kinds: models (Model) and links (Link).
Links include the calculations pages, 2d drawings, more fun, etc.
Anything that we envision as living in a separate HTML document from the
viewer.

Given this, the steps to create a collection are, currently:

1. Create a new directory for the collection under src/collections.
2. Add the model files and HTML files.
3. Add the <collection>.ts file (the name of the file is not technically
   important, but there should probably only be one per collection).
4. Add <collection>.ts as an entry point in webpack.common.js.
5. Add a new HtmlWebpackPlugin entry that uses the new collection index
   chunk and outputs to <collection>/index.html using the model-page.ejs
   template.

Step 3 is almost certain to be the most error-prone. The <collection>.ts
file should look something like

```
import { App, Model, Link } from '../../app'
import pathToModel from './model.gltf'
import pathToCalcs from './calculations.html'

const app = new App({
    title: 'A new collection',
    groups: [
        {
            name: 'a group!',
            items: [
                new Model({name: '3d model', path: pathToModel}),
                new Link({name: 'Calculations', url: pathToCalcs})
            ]
        }
    ]
})
```

The groups in the sidebar are defined by the `groups` array. The first
Model defined in the list of groups is used as the default model for the
viewer. This may be made configurable; for now I think it's entirely
reasonable to just use the first one.