Restructure to show HTML content on the app page
The viewer container is now a generic "content container", and HTML
content can be loaded into the container similar to loading a model. The
App class is responsible for keeping track of what's loaded, and the
buttons manipulate the state of the app. It might be worth making the
onclick methods proper App methods, to make the logic and ownership a
little more clear.