Allow linking directly to items
Items are now linked to directly using their `linkname`, which at the
moment is their name in the sidebar with whitespace replaced by dashes.
This does require names to be unique across, not just within, groups in
the sidebar. At the moment I don't think this is a problem, and it would
be pretty easy to do something different.

To support this, the buttons in the sidebar have all been replaced by
(actual) anchors pointing to the appropriate location, and the actual
changing of the page state is delegated to a 'hashchange' window event
listener.

This complicated having a "default" model, but I wasn't really a fan of
that anyway; so now the "default" is set by the main page linking to
whatever we want the default to be.