Add responsive display mode
When the rendered screen is in a portrait aspect ratio, the content
container no longer changes width in response to the sidebar, and
instead the sidebar is drawn on top of the content. I think this makes
more sense than seeing the calcs/model/etc. squished into a weird narrow
column when on mobile.

When the device is in a landscape aspect ratio, the behavior is the same
as before.