Point to the directory to solve a dev server issue
It seems like a lot of servers redirect you to 'double-angle/' if you
specify going to 'double-angle' and end up at an index.html file;
webpack-dev-server does *not* do this, and so fails to load any content
that is specified by a path relative to the current directory (so in our
case, images).

Weird.