Improve edge rendering
EdgesGeometry accepts a parameter that specifies the angle between faces
that results in an edge being drawn. Setting that (much) higher than the
default removes the undesired edges.

The only problem is that cylinders don't have an edge drawn on the side,
but this can be worked around by making sure that any cylinders are a
different color than anything behind them.