Refactor sidebar items and allow markdown as input
Instead of messing about with App's internals, sidebar items are only
passed the contentContainer div, and return their created content. This
also allows for simplifying the handling of HtmlContent; instead of
using file-loader and then fetch'ing the contents, html-loader is used
instead to just directly load the content. The preprocessor option of
html-loader is used to process pandoc markdown into HTML. This adds an
external dependency on pandoc, which I'm not thrilled by but pandoc
works better than anything else AFAICT.