Change paragraph font and rework CSS inclusion
Instead of having webpack use JS to inject the CSS, it makes more sense
to just be sensible and include them as normal stylesheets.

Font is changed from Tinos to STIX 2 Text. There doesn't appear to be a
CDN for STIX 2, so we'll ship it ourselves. Requires file-handler hooks
for *.woff2 files.