Fix usage of \rm macro
\rm doesn't actually take arguments -- it's a switch that turns on
roman (upright) text for the current block (or w/e the term is). So
for the most part it worked "correctly" here, since it was usually
for variable substitution, when all the variables were gone and the
only text is the units, which should be upright. But there were a
handful of equations that still wanted to have italic text, which
\rm plowed through.